refactor: abstract the error handler
